Working on Design System, Filter and Sort with Entities

+++ app/src/main/java/com/suonk/notepad_plus/ui/note/list/NotesListViewState.kt
@@ -7,7 +7,7 @@ data class NotesListViewState(
     val title: String,
     val content: String,
     val date: String,
-    val color: Long,
+    val color: Int,
     val onItemNoteClicked: EquatableCallback,
     val onDeleteNoteClicked: EquatableCallback,
     val actions: List<NotesListActionViewState>,
